Creating an access_controller model
Create an AccessController model that will allow developers to easily
call the cassia api. These require some changes to the current requests
so that we are passing in the AccessController object. We also are
adding the concept of response handlers which will handle the responses
from the API and any logic of setting values on the access controllers.